// keyStore/server/src/generated/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Json: any;
};


export type Mutation = {
  __typename?: 'Mutation';
  exchangeToken: ExchangeTokenResponse;
  createEntry: CreateEntryResponse;
  importEntry: ImportEntryResponse;
  removeEntry: RemoveEntryResponse;
};


export type MutationExchangeTokenArgs = {
  jwt: Scalars['String'];
};


export type MutationCreateEntryArgs = {
  publicKey?: Maybe<Scalars['String']>;
  entryContent: Scalars['Json'];
};


export type MutationImportEntryArgs = {
  entryHash: Scalars['String'];
  name: Scalars['String'];
  overwrite?: Maybe<Scalars['Boolean']>;
};


export type MutationRemoveEntryArgs = {
  name: Scalars['String'];
};

export type Query = {
  __typename?: 'Query';
  indexEntry?: Maybe<Entry>;
  findEntryByHash?: Maybe<Entry>;
  findEntryByHashCleartext?: Maybe<Entry>;
  version?: Maybe<Version>;
};


export type QueryFindEntryByHashArgs = {
  hash: Scalars['String'];
};


export type QueryFindEntryByHashCleartextArgs = {
  hash: Scalars['String'];
};

export type Entry = {
  __typename?: 'Entry';
  creatorFingerPrint: Scalars['String'];
  ownerFingerPrint: Scalars['String'];
  entryHash: Scalars['String'];
  content: Scalars['Json'];
};

export type ActionResponse = {
  success: Scalars['Boolean'];
  errorMessage?: Maybe<Scalars['String']>;
};

export type ExchangeTokenResponse = ActionResponse & {
  __typename?: 'ExchangeTokenResponse';
  success: Scalars['Boolean'];
  errorMessage?: Maybe<Scalars['String']>;
};

export type CreateEntryResponse = ActionResponse & {
  __typename?: 'CreateEntryResponse';
  success: Scalars['Boolean'];
  errorMessage?: Maybe<Scalars['String']>;
  entryHash?: Maybe<Scalars['String']>;
};

export type RemoveEntryResponse = ActionResponse & {
  __typename?: 'RemoveEntryResponse';
  success: Scalars['Boolean'];
  errorMessage?: Maybe<Scalars['String']>;
  entryHash?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
};

export type ImportEntryResponse = ActionResponse & {
  __typename?: 'ImportEntryResponse';
  success: Scalars['Boolean'];
  errorMessage?: Maybe<Scalars['String']>;
  entryHash?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
};

export type Version = {
  __typename?: 'Version';
  major: Scalars['Int'];
  minor: Scalars['Int'];
  revision: Scalars['Int'];
};

export type WithIndex<TObject> = TObject & Record<string, any>;
export type ResolversObject<TObject> = WithIndex<TObject>;

export type ResolverTypeWrapper<T> = Promise<T> | T;


export type LegacyStitchingResolver<TResult, TParent, TContext, TArgs> = {
  fragment: string;
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};

export type NewStitchingResolver<TResult, TParent, TContext, TArgs> = {
  selectionSet: string;
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type StitchingResolver<TResult, TParent, TContext, TArgs> = LegacyStitchingResolver<TResult, TParent, TContext, TArgs> | NewStitchingResolver<TResult, TParent, TContext, TArgs>;
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | StitchingResolver<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type ResolversTypes = ResolversObject<{
  Json: ResolverTypeWrapper<Scalars['Json']>;
  Mutation: ResolverTypeWrapper<{}>;
  String: ResolverTypeWrapper<Scalars['String']>;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
  Query: ResolverTypeWrapper<{}>;
  Entry: ResolverTypeWrapper<Entry>;
  ActionResponse: ResolversTypes['ExchangeTokenResponse'] | ResolversTypes['CreateEntryResponse'] | ResolversTypes['RemoveEntryResponse'] | ResolversTypes['ImportEntryResponse'];
  ExchangeTokenResponse: ResolverTypeWrapper<ExchangeTokenResponse>;
  CreateEntryResponse: ResolverTypeWrapper<CreateEntryResponse>;
  RemoveEntryResponse: ResolverTypeWrapper<RemoveEntryResponse>;
  ImportEntryResponse: ResolverTypeWrapper<ImportEntryResponse>;
  Version: ResolverTypeWrapper<Version>;
  Int: ResolverTypeWrapper<Scalars['Int']>;
}>;

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = ResolversObject<{
  Json: Scalars['Json'];
  Mutation: {};
  String: Scalars['String'];
  Boolean: Scalars['Boolean'];
  Query: {};
  Entry: Entry;
  ActionResponse: ResolversParentTypes['ExchangeTokenResponse'] | ResolversParentTypes['CreateEntryResponse'] | ResolversParentTypes['RemoveEntryResponse'] | ResolversParentTypes['ImportEntryResponse'];
  ExchangeTokenResponse: ExchangeTokenResponse;
  CreateEntryResponse: CreateEntryResponse;
  RemoveEntryResponse: RemoveEntryResponse;
  ImportEntryResponse: ImportEntryResponse;
  Version: Version;
  Int: Scalars['Int'];
}>;

export interface JsonScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['Json'], any> {
  name: 'Json';
}

export type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = ResolversObject<{
  exchangeToken?: Resolver<ResolversTypes['ExchangeTokenResponse'], ParentType, ContextType, RequireFields<MutationExchangeTokenArgs, 'jwt'>>;
  createEntry?: Resolver<ResolversTypes['CreateEntryResponse'], ParentType, ContextType, RequireFields<MutationCreateEntryArgs, 'entryContent'>>;
  importEntry?: Resolver<ResolversTypes['ImportEntryResponse'], ParentType, ContextType, RequireFields<MutationImportEntryArgs, 'entryHash' | 'name'>>;
  removeEntry?: Resolver<ResolversTypes['RemoveEntryResponse'], ParentType, ContextType, RequireFields<MutationRemoveEntryArgs, 'name'>>;
}>;

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = ResolversObject<{
  indexEntry?: Resolver<Maybe<ResolversTypes['Entry']>, ParentType, ContextType>;
  findEntryByHash?: Resolver<Maybe<ResolversTypes['Entry']>, ParentType, ContextType, RequireFields<QueryFindEntryByHashArgs, 'hash'>>;
  findEntryByHashCleartext?: Resolver<Maybe<ResolversTypes['Entry']>, ParentType, ContextType, RequireFields<QueryFindEntryByHashCleartextArgs, 'hash'>>;
  version?: Resolver<Maybe<ResolversTypes['Version']>, ParentType, ContextType>;
}>;